Creates a custom gradient colormap
Sample usage:

This is the call to the list_by_values function. Input parameters are as many ranges
as desired. In this example, there are 3 ranges, one starting from #00FFFF for a value
of 25 and going to #0000FF for a value of 29.

First, create an instance of the gradient class that stores the data used to make the
color map. Input parameters are as many ranges as desired.
obj = gradient([['#00FFFF',25.0],['#0000FF',29.0]],
               [['#0000FF',29.0],['#0000AA',32.0]],
               [['#0000AA',32.0],['#FF00FF',38.0]])

Create a range of levels going from 25 to 39 with an increment of 0.5:
clevs = np.arange(25.0,39.0,0.5)

To retrieve the gradient colormap matching to the range of levels, use the get_cmap function:
cmap = obj.get_cmap(clevs)